fixes and cleaning
