+++ ydb/core/grpc_services/rpc_create_table.cpp
@@ -109,7 +109,6 @@ private:
         tableDesc->SetName(tableName);
 
         auto schema = tableDesc->MutableSchema();
-        schema->SetEngine(NKikimrSchemeOp::EColumnTableEngine::COLUMN_ENGINE_REPLACING_TIMESERIES);
 
         TString error;
         if (!FillColumnDescription(*tableDesc, req.columns(), code, error)) {
