+++ ydb/core/kqp/ut/olap/helpers/typed_local.cpp
@@ -13,7 +13,6 @@ TString TTypedLocalHelper::GetTestTableSchema() const {
             Columns { Name: "pk_int" Type: "Int64" NotNull: true }
             Columns { Name: "ts" Type: "Timestamp" }
             KeyColumnNames: "pk_int"
-            Engine: COLUMN_ENGINE_REPLACING_TIMESERIES
         )";
     return result;
 }
@@ -34,7 +33,6 @@ TString TTypedLocalHelper::GetMultiColumnTestTableSchema(ui32 reps) const {
     }
     result += R"(
             KeyColumnNames: "pk_int"
-            Engine: COLUMN_ENGINE_REPLACING_TIMESERIES
     )";
     return result;
 }
