+++ ydb/core/tx/columnshard/engines/scheme/objects_cache.cpp
@@ -1,0 +1,5 @@
+#include "objects_cache.h"
+
+namespace NKikimr::NOlap {
+
+}   // namespace NKikimr::NOlap
