+++ ydb/core/tx/schemeshard/olap/schema/update.h
@@ -10,7 +10,6 @@ namespace NKikimr::NSchemeShard {
         YDB_READONLY_DEF(TOlapColumnsUpdate, Columns);
         YDB_READONLY_DEF(TOlapIndexesUpdate, Indexes);
         YDB_READONLY_DEF(TOlapOptionsUpdate, Options);
-        YDB_READONLY_OPT(NKikimrSchemeOp::EColumnTableEngine, Engine);
     public:
         bool Parse(const NKikimrSchemeOp::TColumnTableSchema& tableSchema, IErrorCollector& errors, bool allowNullKeys = false);
         bool Parse(const NKikimrSchemeOp::TAlterColumnTableSchema& alterRequest, IErrorCollector& errors);
